# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitoring Karyawan API",
        "description": "API Documentation for Monitoring Karyawan",
        "contact": {
            "email": "admin@monitoringkaryawan.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://monitoringkaryawan.local/",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/penugasans": {
            "get": {
                "tags": [
                    "Penugasan"
                ],
                "summary": "Get all penugasans",
                "operationId": "f244c661e2c717493bad335050639590",
                "responses": {
                    "200": {
                        "description": "List of all penugasans"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Penugasan"
                ],
                "summary": "Create a new penugasan (assign task)",
                "operationId": "981344cdfcc27b730f3ec917cd28d95b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assignable_type",
                                    "assignable_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "assignable_type": {
                                        "type": "string",
                                        "enum": [
                                            "HO",
                                            "Canvasing"
                                        ]
                                    },
                                    "assignable_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Penugasan created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Assignable record not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penugasans/{uuid}/status": {
            "post": {
                "tags": [
                    "Penugasan"
                ],
                "summary": "Update status of a penugasan",
                "operationId": "530356c2abbe0bc6f011174b218da033",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "status_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "foto[]": {
                                        "description": "Max 3 images (jpeg, png, jpg)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "404": {
                        "description": "Penugasan not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penugasans/{uuid}": {
            "delete": {
                "tags": [
                    "Penugasan"
                ],
                "summary": "Delete (cancel) a penugasan",
                "operationId": "3e7a5c52279f9a9eaf15b039d0abdbba",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Penugasan deleted"
                    },
                    "404": {
                        "description": "Penugasan not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}